smartAssistance/assist_requirements.txt
langchain
langchain_community
langchain-groq
langgraph
python-dotenv
grandalf
arxiv
wikipedia
winshell
tempfile
selenium
